Create tests

+++ src/main/java/io/github/hildi/can/service/StandardJavaSerializationService.java
@@ -1,5 +1,6 @@
 package io.github.hildi.can.service;
 
+import io.github.hildi.can.exceptions.CannotReadFileException;
 import io.github.hildi.can.model.User;
 
 import java.io.*;
@@ -10,16 +11,24 @@ import java.io.*;
 public class StandardJavaSerializationService implements SerializationService {
 
     @Override
-    public void serialize(User user, File file) throws IOException {
+    public void serialize(User user, File file) {
+
+        if (!(file.exists() || file.canWrite())) { return; }
         try (ObjectOutputStream out = new ObjectOutputStream(new FileOutputStream(file))){
             out.writeObject(user);
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 
     @Override
-    public User deserialize(File file) throws IOException, ClassNotFoundException {
+    public User deserialize(File file) {
         try (ObjectInputStream in = new ObjectInputStream(new FileInputStream(file))) {
             return (User) in.readObject();
+        } catch (EOFException ignored) {
+        } catch (IOException | ClassNotFoundException e) {
+            e.printStackTrace();
         }
+        throw new CannotReadFileException("We cannot read bytes of array at this files.");
     }
 }
